Refactored Configuration.cs to a separate project
This change is needed because Log() requires parameters from
Configuration.cs
